backend/requirements.txt
# ============================================================
# Hardware
# ============================================================
psutil>=5.9
GPUtil>=1.4
nvidia-ml-py>=12.0
onnxruntime>=1.17
pyrsmi>=1.0

# ============================================================
# Voice / Speech
# ============================================================
faster-whisper>=1.0
nvidia-cublas-cu12>=12.0
nvidia-cudnn-cu12>=9.0
sounddevice>=0.4
soundfile>=0.12

# ============================================================
# LLM / AI
# ============================================================
huggingface_hub>=0.23
ollama>=0.4
httpx>=0.27

# ============================================================
# Web / API Core
# ============================================================

# ============================================================
# Configuration / Settings
# ============================================================
PyYAML>=6.0

# ============================================================
# Storage / Database (SQLite)
# ============================================================

# ============================================================
# Caching / Lightweight State
# ============================================================

# ============================================================
# Search / Scraping
# ============================================================

# ============================================================
# Networking / HTTP
# ============================================================

# ============================================================
# Security / Auth
# ============================================================

# ============================================================
# Testing
# ============================================================
pytest>=8.0

# ============================================================
# Utilities
# ============================================================
